render_final/requirements.txt
Flask
Werkzeug
openai
python-dotenv
Flask-Session
Flask-SQLAlchemy
psycopg2-binary
uuid
gunicorn